airflow/requirements.txt
pandas
tqdm
python-dotenv
requests
psycopg2-binary
dbt-core
dbt-postgres